[REFACTORISATION] - Split parsing modules into multiple module
[+] Reactor parsing modules
[+] Refactor utils modules
[+] Refactor tests
