+++ src/tokenizer.rs
@@ -1,8 +1,9 @@
+use regex::Regex;
+
 use crate::{
     elements::{Category, Elements},
-    keyword::{Keyword, Manager},
-    parsing::{is_crc32, is_resolution, ordinals_to_nb},
-    utils::{is_digit, split_by_delimiter},
+    keyword::{Keyword, Manager}, parsing::number::{is_crc32, is_resolution, is_digit, ordinals_to_nb}, split::split_by_delimiter,
+
 };
 
 pub fn tokenize(string_to_tokenize: &str, delimiter: &Vec<char>) -> Vec<Token> {
@@ -166,6 +167,7 @@ impl Token {
     pub fn get_tokens(&self) -> Vec<SubToken>{
         self.tokens.to_owned()
     }
+
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -193,6 +195,56 @@ impl SubToken {
 
     pub fn value(&self) -> String{
         self.value.clone()
+    }
+
+    pub fn match_episode_patern(&self, found_elements: &mut Elements) -> bool {
+        let tested_value = self.value();
+
+        // Multi episode matching test
+        let multiple_ep_regex = Regex::new(r"(?P<ep_1>\d{1,4})(?:[vV](?P<version_1>\d))?[-~&+](?P<ep_2>\d{1,4})(?:[vV](?P<version_2>\d))?$").unwrap();
+        if multiple_ep_regex.is_match(&tested_value) {
+            let multiple_ep_captures = multiple_ep_regex.captures(&tested_value).unwrap();
+            let ep_1 = multiple_ep_captures.name("ep_1").unwrap().as_str();
+            let ep_2 = multiple_ep_captures.name("ep_2").unwrap().as_str();
+            if ep_1.parse::<i32>().unwrap() > ep_2.parse::<i32>().unwrap() {
+                return false;
+            }
+            found_elements.add(Category::EpisodeNumber, ep_1);
+            found_elements.add(Category::EpisodeNumberAlt, ep_2);
+
+            if let Some(v1) = multiple_ep_captures.name("version_1") {
+                found_elements.add(Category::ReleaseVersion, v1.as_str());
+            }
+            if let Some(v2) = multiple_ep_captures.name("version_2") {
+                found_elements.add(Category::ReleaseVersion, v2.as_str());
+            }
+            return true;
+        }
+
+        // Saeson and episode
+        let season_episode_regex = Regex::new(r"S?(?P<season_1>\d{1,2})(?:-S?(?P<season_2>\d{1,2}))?(?:x|[ ._-x]?E)(?P<ep_1>\d{1,4})(?:-E?(?P<ep_2>\d{1,4}))?(?:[vV](?P<version>\d))?$").unwrap();
+        if season_episode_regex.is_match(&tested_value) {
+            let season_ep_captures = season_episode_regex.captures(&tested_value).unwrap();
+            let season_1 = season_ep_captures.name("season_1").unwrap().as_str();
+            if season_1.parse::<i32>().unwrap() == 0 {
+                return false;
+            }
+            found_elements.add(Category::AnimeSeason, season_1);
+            if let Some(season_2) = season_ep_captures.name("season_2") {
+                found_elements.add(Category::AnimeSeason, season_2.as_str());
+            }
+            let ep_1 = season_ep_captures.name("ep_1").unwrap().as_str();
+            found_elements.add(Category::EpisodeNumber, ep_1);
+            if let Some(ep_2) = season_ep_captures.name("ep_2") {
+                found_elements.add(Category::AnimeSeason, ep_2.as_str());
+            }
+            if let Some(version) = season_ep_captures.name("version") {
+                found_elements.add(Category::AnimeSeason, version.as_str());
+            }
+            return true;
+        }
+
+        false
     }
 }
