+++ src/utils.rs
@@ -2,1796 +2,6 @@ use regex::Regex;
 use std::path::Path;
 use unicode_normalization::UnicodeNormalization;
 
-fn is_valid_extension(tested_extension: &str) -> bool {
-    let valid_extension = vec![
-        ".php".to_string(),
-        ".html".to_string(),
-        ".txt".to_string(),
-        ".htm".to_string(),
-        ".aspx".to_string(),
-        ".asp".to_string(),
-        ".js".to_string(),
-        ".css".to_string(),
-        ".pgsql.txt".to_string(),
-        ".mysql.txt".to_string(),
-        ".pdf".to_string(),
-        ".cgi".to_string(),
-        ".inc".to_string(),
-        ".gif".to_string(),
-        ".jpg".to_string(),
-        ".swf".to_string(),
-        ".xml".to_string(),
-        ".cfm".to_string(),
-        ".xhtml".to_string(),
-        ".wmv".to_string(),
-        ".zip".to_string(),
-        ".axd".to_string(),
-        ".gz".to_string(),
-        ".png".to_string(),
-        ".doc".to_string(),
-        ".shtml".to_string(),
-        ".jsp".to_string(),
-        ".ico".to_string(),
-        ".exe".to_string(),
-        ".csi".to_string(),
-        ".inc.php".to_string(),
-        ".config".to_string(),
-        ".jpeg".to_string(),
-        ".ashx".to_string(),
-        ".log".to_string(),
-        ".xls".to_string(),
-        ".old".to_string(),
-        ".mp3".to_string(),
-        ".com".to_string(),
-        ".tar".to_string(),
-        ".ini".to_string(),
-        ".asa".to_string(),
-        ".tgz".to_string(),
-        ".PDF".to_string(),
-        ".flv".to_string(),
-        ".php3".to_string(),
-        ".bak".to_string(),
-        ".rar".to_string(),
-        ".asmx".to_string(),
-        ".xlsx".to_string(),
-        ".page".to_string(),
-        ".phtml".to_string(),
-        ".dll".to_string(),
-        ".JPG".to_string(),
-        ".asax".to_string(),
-        ".msg".to_string(),
-        ".pl".to_string(),
-        ".GIF".to_string(),
-        ".ZIP".to_string(),
-        ".csv".to_string(),
-        ".css.aspx".to_string(),
-        ".JPEG".to_string(),
-        ".ppt".to_string(),
-        ".nsf".to_string(),
-        ".Pdf".to_string(),
-        ".Gif".to_string(),
-        ".bmp".to_string(),
-        ".sql".to_string(),
-        ".Jpeg".to_string(),
-        ".Jpg".to_string(),
-        ".xml.gz".to_string(),
-        ".Zip".to_string(),
-        ".new".to_string(),
-        ".avi".to_string(),
-        ".psd".to_string(),
-        ".rss".to_string(),
-        ".wav".to_string(),
-        ".action".to_string(),
-        ".db".to_string(),
-        ".dat".to_string(),
-        ".do".to_string(),
-        ".xsl".to_string(),
-        ".class".to_string(),
-        ".mdb".to_string(),
-        ".include".to_string(),
-        ".cs".to_string(),
-        ".class.php".to_string(),
-        ".htc".to_string(),
-        ".mov".to_string(),
-        ".tpl".to_string(),
-        ".js.php".to_string(),
-        ".mysql-connect".to_string(),
-        ".mpg".to_string(),
-        ".rdf".to_string(),
-        ".rtf".to_string(),
-        ".ascx".to_string(),
-        ".mvc".to_string(),
-        ".1.0".to_string(),
-        ".files".to_string(),
-        ".master".to_string(),
-        ".jar".to_string(),
-        ".vb".to_string(),
-        ".mp4".to_string(),
-        ".local.php".to_string(),
-        ".fla".to_string(),
-        ".require".to_string(),
-        ".de".to_string(),
-        ".docx".to_string(),
-        ".php5".to_string(),
-        ".wci".to_string(),
-        ".readme".to_string(),
-        ".cfg".to_string(),
-        ".aspx.cs".to_string(),
-        ".cfc".to_string(),
-        ".dwt".to_string(),
-        ".ru".to_string(),
-        ".LCK".to_string(),
-        ".Config".to_string(),
-        ".gif_var_DE".to_string(),
-        ".html_var_DE".to_string(),
-        ".net".to_string(),
-        ".ttf".to_string(),
-        ".HTM".to_string(),
-        ".X-AOM".to_string(),
-        ".jhtml".to_string(),
-        ".mpeg".to_string(),
-        ".ASP".to_string(),
-        ".LOG".to_string(),
-        ".X-FANCYCAT".to_string(),
-        ".php4".to_string(),
-        ".readme_var_DE".to_string(),
-        ".vcf".to_string(),
-        ".X-RMA".to_string(),
-        ".X-AFFILIATE".to_string(),
-        ".X-OFFERS".to_string(),
-        ".X-AFFILIATE_var_DE".to_string(),
-        ".X-AOM_var_DE".to_string(),
-        ".X-FANCYCAT_var_DE".to_string(),
-        ".X-FCOMP".to_string(),
-        ".X-FCOMP_var_DE".to_string(),
-        ".X-GIFTREG".to_string(),
-        ".X-GIFTREG_var_DE".to_string(),
-        ".X-MAGNIFIER".to_string(),
-        ".X-MAGNIFIER_var_DE".to_string(),
-        ".X-OFFERS_var_DE".to_string(),
-        ".X-PCONF".to_string(),
-        ".X-PCONF_var_DE".to_string(),
-        ".X-RMA_var_DE".to_string(),
-        ".X-SURVEY".to_string(),
-        ".tif".to_string(),
-        ".dir".to_string(),
-        ".json".to_string(),
-        ".6.9".to_string(),
-        ".Zif".to_string(),
-        ".wma".to_string(),
-        ".mid".to_string(),
-        ".rm".to_string(),
-        ".aspx.vb".to_string(),
-        ".tar.gz".to_string(),
-        ".woa".to_string(),
-        ".main".to_string(),
-        ".ram".to_string(),
-        ".opml".to_string(),
-        ".0.html".to_string(),
-        ".css.php".to_string(),
-        ".feed".to_string(),
-        ".lasso".to_string(),
-        ".6.3".to_string(),
-        ".shtm".to_string(),
-        ".sitemap".to_string(),
-        ".scc".to_string(),
-        ".tmp".to_string(),
-        ".backup".to_string(),
-        ".sln".to_string(),
-        ".org".to_string(),
-        ".conf".to_string(),
-        ".mysql-query".to_string(),
-        ".session-start".to_string(),
-        ".uk".to_string(),
-        ".TXT".to_string(),
-        ".orig".to_string(),
-        ".settings.php".to_string(),
-        ".cab".to_string(),
-        ".kml".to_string(),
-        ".lck".to_string(),
-        ".pps".to_string(),
-        ".require-once".to_string(),
-        ".asx".to_string(),
-        ".bok".to_string(),
-        ".msi".to_string(),
-        ".c".to_string(),
-        ".fcgi".to_string(),
-        ".fopen".to_string(),
-        ".html.".to_string(),
-        ".phpmailer.php".to_string(),
-        ".bin".to_string(),
-        ".htaccess".to_string(),
-        ".info".to_string(),
-        ".java".to_string(),
-        ".jsf".to_string(),
-        ".tmpl".to_string(),
-        ".DOC".to_string(),
-        ".bat".to_string(),
-        ".com.html".to_string(),
-        ".print".to_string(),
-        ".resx".to_string(),
-        ".ics".to_string(),
-        ".php.php".to_string(),
-        ".x".to_string(),
-        ".PNG".to_string(),
-        ".data".to_string(),
-        ".dcr".to_string(),
-        ".enfinity".to_string(),
-        ".html.html".to_string(),
-        ".licx".to_string(),
-        ".mno".to_string(),
-        ".plx".to_string(),
-        ".vm".to_string(),
-        ".5.php".to_string(),
-        ".HTML".to_string(),
-        ".MP3".to_string(),
-        ".config.php".to_string(),
-        ".dwg".to_string(),
-        ".edu".to_string(),
-        ".search".to_string(),
-        ".static".to_string(),
-        ".wws".to_string(),
-        ".6.edu".to_string(),
-        ".OLD".to_string(),
-        ".bz2".to_string(),
-        ".co.uk".to_string(),
-        ".ece".to_string(),
-        ".epc".to_string(),
-        ".getimagesize".to_string(),
-        ".ice".to_string(),
-        ".it_Backup_Giornaliero".to_string(),
-        ".it_Backup_Settimanale".to_string(),
-        ".jspa".to_string(),
-        ".lst".to_string(),
-        ".php-dist".to_string(),
-        ".svc".to_string(),
-        ".vbs".to_string(),
-        ".1.html".to_string(),
-        ".30-i486".to_string(),
-        ".ai".to_string(),
-        ".cur".to_string(),
-        ".dmg".to_string(),
-        ".img".to_string(),
-        ".inf".to_string(),
-        ".seam".to_string(),
-        ".smtp.php".to_string(),
-        ".1-bin-Linux-2.0.30-i486".to_string(),
-        ".7z".to_string(),
-        ".ajax".to_string(),
-        ".cfm.cfm".to_string(),
-        ".chm".to_string(),
-        ".csp".to_string(),
-        ".edit".to_string(),
-        ".file".to_string(),
-        ".gif.php".to_string(),
-        ".m3u".to_string(),
-        ".psp".to_string(),
-        ".py".to_string(),
-        ".sh".to_string(),
-        ".test".to_string(),
-        ".zdat".to_string(),
-        ".admin".to_string(),
-        ".captcha.aspx".to_string(),
-        ".dev".to_string(),
-        ".eps".to_string(),
-        ".file-get-contents".to_string(),
-        ".fr".to_string(),
-        ".fsockopen".to_string(),
-        ".list".to_string(),
-        ".m4v".to_string(),
-        ".min.js".to_string(),
-        ".new.html".to_string(),
-        ".p".to_string(),
-        ".store".to_string(),
-        ".webinfo".to_string(),
-        ".xml.php".to_string(),
-        ".BAK".to_string(),
-        ".htm.".to_string(),
-        ".php.bak".to_string(),
-        ".bk".to_string(),
-        ".bsp".to_string(),
-        ".cms".to_string(),
-        ".csshandler.ashx".to_string(),
-        ".d".to_string(),
-        ".html,".to_string(),
-        ".htmll".to_string(),
-        ".idx".to_string(),
-        ".images".to_string(),
-        ".jad".to_string(),
-        ".master.cs".to_string(),
-        ".prev_next".to_string(),
-        ".ssf".to_string(),
-        ".stm".to_string(),
-        ".txt.gz".to_string(),
-        ".Web.UI.WebResource.axd".to_string(),
-        ".as".to_string(),
-        ".asp.asp".to_string(),
-        ".au".to_string(),
-        ".cnf".to_string(),
-        ".dhtml".to_string(),
-        ".enu".to_string(),
-        ".html.old".to_string(),
-        ".include-once".to_string(),
-        ".lock".to_string(),
-        ".m".to_string(),
-        ".mysql-select-db".to_string(),
-        ".phps".to_string(),
-        ".pm".to_string(),
-        ".pptx".to_string(),
-        ".sav".to_string(),
-        ".sendtoafriendform".to_string(),
-        ".ssi".to_string(),
-        ".suo".to_string(),
-        ".vbproj".to_string(),
-        ".wml".to_string(),
-        ".xsd".to_string(),
-        ".26.13.391N35.50.38.816".to_string(),
-        ".26.24.165N35.50.24.134".to_string(),
-        ".26.56.247N35.52.03.605".to_string(),
-        ".27.02.940N35.49.56.075".to_string(),
-        ".27.15.919N35.52.04.300".to_string(),
-        ".27.29.262N35.47.15.083".to_string(),
-        ".3gp".to_string(),
-        ".40.00.573N35.42.57.445".to_string(),
-        ".43.58.040N35.38.35.826".to_string(),
-        ".44.04.344N35.38.35.077".to_string(),
-        ".44.08.714N35.39.08.499".to_string(),
-        ".44.10.892N35.38.49.246".to_string(),
-        ".44.27.243N35.41.29.367".to_string(),
-        ".44.29.976N35.37.51.790".to_string(),
-        ".44.32.445N35.36.10.206".to_string(),
-        ".44.34.800N35.38.08.156".to_string(),
-        ".44.37.128N35.40.54.403".to_string(),
-        ".44.40.556N35.40.53.025".to_string(),
-        ".44.45.013N35.38.36.211".to_string(),
-        ".44.46.104N35.38.22.970".to_string(),
-        ".44.48.130N35.38.25.969".to_string(),
-        ".44.52.162N35.38.50.456".to_string(),
-        ".44.58.315N35.38.53.455".to_string(),
-        ".45.01.562N35.38.38.778".to_string(),
-        ".45.04.359N35.38.39.112".to_string(),
-        ".45.06.789N35.38.22.556".to_string(),
-        ".45.10.717N35.38.41.989".to_string(),
-        ".ASPX".to_string(),
-        ".JS".to_string(),
-        ".PHP".to_string(),
-        ".array-keys".to_string(),
-        ".atom".to_string(),
-        ".award".to_string(),
-        ".bkp".to_string(),
-        ".crt".to_string(),
-        ".default".to_string(),
-        ".eml".to_string(),
-        ".epl".to_string(),
-        ".fancybox".to_string(),
-        ".fil".to_string(),
-        ".geo".to_string(),
-        ".h".to_string(),
-        ".hmtl".to_string(),
-        ".html.bak".to_string(),
-        ".ida".to_string(),
-        ".implode".to_string(),
-        ".index.php".to_string(),
-        ".iso".to_string(),
-        ".kmz".to_string(),
-        ".mysql-pconnect".to_string(),
-        ".php.old".to_string(),
-        ".php.txt".to_string(),
-        ".rec".to_string(),
-        ".storefront".to_string(),
-        ".taf".to_string(),
-        ".war".to_string(),
-        ".xslt".to_string(),
-        ".1.6".to_string(),
-        ".2a".to_string(),
-        ".8.1".to_string(),
-        ".CSS".to_string(),
-        ".NSF".to_string(),
-        ".Sponsors".to_string(),
-        ".a".to_string(),
-        ".aquery".to_string(),
-        ".ascx.cs".to_string(),
-        ".cat".to_string(),
-        ".contrib".to_string(),
-        ".ds".to_string(),
-        ".dwf".to_string(),
-        ".film".to_string(),
-        ".g".to_string(),
-        ".go".to_string(),
-        ".googlebook".to_string(),
-        ".gpx".to_string(),
-        ".hotelName".to_string(),
-        ".htm.htm".to_string(),
-        ".ihtml".to_string(),
-        ".in-array".to_string(),
-        ".index".to_string(),
-        ".ini.php".to_string(),
-        ".layer".to_string(),
-        ".maninfo".to_string(),
-        ".odt".to_string(),
-        ".price".to_string(),
-        ".randomhouse".to_string(),
-        ".read".to_string(),
-        ".ru-tov.html".to_string(),
-        ".s7".to_string(),
-        ".sample".to_string(),
-        ".sit".to_string(),
-        ".src".to_string(),
-        ".tpl.php".to_string(),
-        ".trck".to_string(),
-        ".uguide".to_string(),
-        ".vorteil".to_string(),
-        ".wbp".to_string(),
-        ".2.html".to_string(),
-        ".AVI".to_string(),
-        ".Asp".to_string(),
-        ".EXE".to_string(),
-        ".WMV".to_string(),
-        ".asax.vb".to_string(),
-        ".aspx.aspx".to_string(),
-        ".btr".to_string(),
-        ".cer".to_string(),
-        ".common.php".to_string(),
-        ".de.html".to_string(),
-        ".html‎".to_string(),
-        ".jbf".to_string(),
-        ".lbi".to_string(),
-        ".lib.php".to_string(),
-        ".lnk".to_string(),
-        ".login".to_string(),
-        ".login.php".to_string(),
-        ".mhtml".to_string(),
-        ".mpl".to_string(),
-        ".mso".to_string(),
-        ".mysql-result".to_string(),
-        ".original".to_string(),
-        ".pgp".to_string(),
-        ".ph".to_string(),
-        ".php.".to_string(),
-        ".preview".to_string(),
-        ".preview-content.php".to_string(),
-        ".search.htm".to_string(),
-        ".site".to_string(),
-        ".text".to_string(),
-        ".view".to_string(),
-        ".3.html".to_string(),
-        ".4.html".to_string(),
-        ".5.html".to_string(),
-        ".ICO".to_string(),
-        ".Web".to_string(),
-        ".XLS".to_string(),
-        ".action2".to_string(),
-        ".asc".to_string(),
-        ".asp.bak".to_string(),
-        ".aspx.resx".to_string(),
-        ".browse".to_string(),
-        ".code".to_string(),
-        ".com_Backup_Giornaliero".to_string(),
-        ".com_Backup_Settimanale".to_string(),
-        ".csproj".to_string(),
-        ".dtd".to_string(),
-        ".en.html".to_string(),
-        ".ep".to_string(),
-        ".eu".to_string(),
-        ".form".to_string(),
-        ".html1".to_string(),
-        ".inc.asp".to_string(),
-        ".index.html".to_string(),
-        ".it".to_string(),
-        ".nl".to_string(),
-        ".ogg".to_string(),
-        ".old.php".to_string(),
-        ".old2".to_string(),
-        ".opendir".to_string(),
-        ".out".to_string(),
-        ".pgt".to_string(),
-        ".php,".to_string(),
-        ".php‎".to_string(),
-        ".po".to_string(),
-        ".prt".to_string(),
-        ".query".to_string(),
-        ".rb".to_string(),
-        ".rhtml".to_string(),
-        ".ru.html".to_string(),
-        ".save".to_string(),
-        ".search.php".to_string(),
-        ".t".to_string(),
-        ".wsdl".to_string(),
-        ".0-to1.2.php".to_string(),
-        ".CFM".to_string(),
-        ".MOV".to_string(),
-        ".MPEG".to_string(),
-        ".Master".to_string(),
-        ".PPT".to_string(),
-        ".TTF".to_string(),
-        ".Templates".to_string(),
-        ".XML".to_string(),
-        ".adp".to_string(),
-        ".ajax.php".to_string(),
-        ".apsx".to_string(),
-        ".asf".to_string(),
-        ".bck".to_string(),
-        ".bu".to_string(),
-        ".calendar".to_string(),
-        ".captcha".to_string(),
-        ".cart".to_string(),
-        ".com.crt".to_string(),
-        ".core".to_string(),
-        ".dict.php".to_string(),
-        ".dot".to_string(),
-        ".egov".to_string(),
-        ".en.php".to_string(),
-        ".eot".to_string(),
-        ".errors".to_string(),
-        ".f4v".to_string(),
-        ".fr.html".to_string(),
-        ".git".to_string(),
-        ".ht".to_string(),
-        ".hta".to_string(),
-        ".html.LCK".to_string(),
-        ".html.printable".to_string(),
-        ".ini.sample".to_string(),
-        ".lib".to_string(),
-        ".lic".to_string(),
-        ".map".to_string(),
-        ".master.vb".to_string(),
-        ".mi".to_string(),
-        ".mkdir".to_string(),
-        ".o".to_string(),
-        ".p7b".to_string(),
-        ".pac".to_string(),
-        ".parse.errors".to_string(),
-        ".pd".to_string(),
-        ".pfx".to_string(),
-        ".php2".to_string(),
-        ".php_files".to_string(),
-        ".phtm".to_string(),
-        ".png.php".to_string(),
-        ".portal".to_string(),
-        ".printable".to_string(),
-        ".psql".to_string(),
-        ".pub".to_string(),
-        ".q".to_string(),
-        ".ra".to_string(),
-        ".reg".to_string(),
-        ".restrictor.php".to_string(),
-        ".rpm".to_string(),
-        ".strpos".to_string(),
-        ".tcl".to_string(),
-        ".template".to_string(),
-        ".tiff".to_string(),
-        ".tv".to_string(),
-        ".us".to_string(),
-        ".user".to_string(),
-        ".Controls".to_string(),
-        ".WAV".to_string(),
-        ".acgi".to_string(),
-        ".alt".to_string(),
-        ".array-merge".to_string(),
-        ".back".to_string(),
-        ".call-user-func-array".to_string(),
-        ".cfml".to_string(),
-        ".cmd".to_string(),
-        ".cocomore.txt".to_string(),
-        ".detail".to_string(),
-        ".disabled".to_string(),
-        ".dist.php".to_string(),
-        ".djvu".to_string(),
-        ".dta".to_string(),
-        ".e".to_string(),
-        ".extract".to_string(),
-        ".file-put-contents".to_string(),
-        ".fpl".to_string(),
-        ".framework".to_string(),
-        ".fread".to_string(),
-        ".htm.LCK".to_string(),
-        ".inc.js".to_string(),
-        ".includes".to_string(),
-        ".jp".to_string(),
-        ".jpg.html".to_string(),
-        ".l".to_string(),
-        ".letter".to_string(),
-        ".local".to_string(),
-        ".num".to_string(),
-        ".pem".to_string(),
-        ".php.sample".to_string(),
-        ".php}".to_string(),
-        ".php~".to_string(),
-        ".pot".to_string(),
-        ".preg-match".to_string(),
-        ".process".to_string(),
-        ".ps".to_string(),
-        ".r".to_string(),
-        ".raw".to_string(),
-        ".rc".to_string(),
-        ".s".to_string(),
-        ".search.".to_string(),
-        ".server".to_string(),
-        ".sis".to_string(),
-        ".sql.gz".to_string(),
-        ".squery".to_string(),
-        ".subscribe".to_string(),
-        ".svg".to_string(),
-        ".svn".to_string(),
-        ".thtml".to_string(),
-        ".tpl.html".to_string(),
-        ".ua".to_string(),
-        ".vcs".to_string(),
-        ".xhtm".to_string(),
-        ".xml.asp".to_string(),
-        ".xpi".to_string(),
-        ".A".to_string(),
-        ".PAGE".to_string(),
-        ".SWF".to_string(),
-        ".add".to_string(),
-        ".array-rand".to_string(),
-        ".asax.cs".to_string(),
-        ".asax.resx".to_string(),
-        ".ascx.vb".to_string(),
-        ".aspx,".to_string(),
-        ".aspx.".to_string(),
-        ".awm".to_string(),
-        ".b".to_string(),
-        ".bhtml".to_string(),
-        ".bml".to_string(),
-        ".ca".to_string(),
-        ".cache".to_string(),
-        ".cfg.php".to_string(),
-        ".cn".to_string(),
-        ".cz".to_string(),
-        ".de.txt".to_string(),
-        ".diff".to_string(),
-        ".email".to_string(),
-        ".en".to_string(),
-        ".error".to_string(),
-        ".faces".to_string(),
-        ".filesize".to_string(),
-        ".functions.php".to_string(),
-        ".hml".to_string(),
-        ".hqx".to_string(),
-        ".html,404".to_string(),
-        ".html.php".to_string(),
-        ".htmls".to_string(),
-        ".htx".to_string(),
-        ".i".to_string(),
-        ".idq".to_string(),
-        ".jpe".to_string(),
-        ".js.aspx".to_string(),
-        ".js.gz".to_string(),
-        ".jspf".to_string(),
-        ".load".to_string(),
-        ".media".to_string(),
-        ".mp2".to_string(),
-        ".mspx".to_string(),
-        ".mv".to_string(),
-        ".mysql".to_string(),
-        ".new.php".to_string(),
-        ".ocx".to_string(),
-        ".oui".to_string(),
-        ".outcontrol".to_string(),
-        ".pad".to_string(),
-        ".pages".to_string(),
-        ".pdb".to_string(),
-        ".pdf.".to_string(),
-        ".pnp".to_string(),
-        ".pop_formata_viewer".to_string(),
-        ".popup.php".to_string(),
-        ".popup.pop_formata_viewer".to_string(),
-        ".pvk".to_string(),
-        ".restrictor.log".to_string(),
-        ".results".to_string(),
-        ".run".to_string(),
-        ".scripts".to_string(),
-        ".sdb".to_string(),
-        ".ser".to_string(),
-        ".shop".to_string(),
-        ".sitemap.xml".to_string(),
-        ".smi".to_string(),
-        ".start".to_string(),
-        ".ste".to_string(),
-        ".swf.swf".to_string(),
-        ".templates".to_string(),
-        ".textsearch".to_string(),
-        ".torrent".to_string(),
-        ".unsubscribe".to_string(),
-        ".v".to_string(),
-        ".vbproj.webinfo".to_string(),
-        ".web".to_string(),
-        ".wmf".to_string(),
-        ".wpd".to_string(),
-        ".ws".to_string(),
-        ".xpml".to_string(),
-        ".y".to_string(),
-        ".AdCode".to_string(),
-        ".Aspx".to_string(),
-        ".C.".to_string(),
-        ".COM".to_string(),
-        ".GetMapImage".to_string(),
-        ".Html".to_string(),
-        ".Run.AdCode".to_string(),
-        ".Skins".to_string(),
-        ".Z".to_string(),
-        ".access.login".to_string(),
-        ".ajax.asp".to_string(),
-        ".app".to_string(),
-        ".asd".to_string(),
-        ".asm".to_string(),
-        ".assets".to_string(),
-        ".at".to_string(),
-        ".bad".to_string(),
-        ".bak2".to_string(),
-        ".blog".to_string(),
-        ".casino".to_string(),
-        ".cc".to_string(),
-        ".cdr".to_string(),
-        ".changeLang.php".to_string(),
-        ".children".to_string(),
-        ".com,".to_string(),
-        ".com-redirect".to_string(),
-        ".content".to_string(),
-        ".copy".to_string(),
-        ".count".to_string(),
-        ".cp".to_string(),
-        ".csproj.user".to_string(),
-        ".custom".to_string(),
-        ".dbf".to_string(),
-        ".deb".to_string(),
-        ".delete".to_string(),
-        ".details.php".to_string(),
-        ".dic".to_string(),
-        ".divx".to_string(),
-        ".download".to_string(),
-        ".download.php".to_string(),
-        ".downloadCirRequirements.pdf".to_string(),
-        ".downloadTourkitRequirements.pdf".to_string(),
-        ".emailCirRequirements.php".to_string(),
-        ".emailTourkitForm.php".to_string(),
-        ".emailTourkitNotification.php".to_string(),
-        ".emailTourkitRequirements.php".to_string(),
-        ".epub".to_string(),
-        ".err".to_string(),
-        ".es".to_string(),
-        ".exclude".to_string(),
-        ".filemtime".to_string(),
-        ".fillPurposes2.php".to_string(),
-        ".grp".to_string(),
-        ".home".to_string(),
-        ".htlm".to_string(),
-        ".htm,".to_string(),
-        ".html-".to_string(),
-        ".image".to_string(),
-        ".inc.html".to_string(),
-        ".it.html".to_string(),
-        ".j".to_string(),
-        ".jnlp".to_string(),
-        ".js.asp".to_string(),
-        ".js2".to_string(),
-        ".jspx".to_string(),
-        ".lang-en.php".to_string(),
-        ".link".to_string(),
-        ".listevents".to_string(),
-        ".log.0".to_string(),
-        ".mbox".to_string(),
-        ".mc_id".to_string(),
-        ".menu.php".to_string(),
-        ".mgi".to_string(),
-        ".mod".to_string(),
-        ".net.html".to_string(),
-        ".news".to_string(),
-        ".none".to_string(),
-        ".off".to_string(),
-        ".p3p".to_string(),
-        ".php.htm".to_string(),
-        ".php.static".to_string(),
-        ".php1".to_string(),
-        ".phpp".to_string(),
-        ".pop3.php".to_string(),
-        ".pop_3D_viewer".to_string(),
-        ".popup.pop_3D_viewer".to_string(),
-        ".prep".to_string(),
-        ".prg".to_string(),
-        ".print.html".to_string(),
-        ".print.php".to_string(),
-        ".product_details".to_string(),
-        ".pwd".to_string(),
-        ".pyc".to_string(),
-        ".red".to_string(),
-        ".registration".to_string(),
-        ".requirementsFeesTable.php".to_string(),
-        ".roshani-gunewardene.com".to_string(),
-        ".se".to_string(),
-        ".sea".to_string(),
-        ".sema".to_string(),
-        ".session".to_string(),
-        ".setup".to_string(),
-        ".simplexml-load-file".to_string(),
-        ".sitx".to_string(),
-        ".smil".to_string(),
-        ".srv".to_string(),
-        ".swi".to_string(),
-        ".swp".to_string(),
-        ".sxw".to_string(),
-        ".tar.bz2".to_string(),
-        ".tem".to_string(),
-        ".temp".to_string(),
-        ".template.php".to_string(),
-        ".top".to_string(),
-        ".txt.php".to_string(),
-        ".types".to_string(),
-        ".unlink".to_string(),
-        ".url".to_string(),
-        ".userLoginPopup.php".to_string(),
-        ".visaPopup.php".to_string(),
-        ".visaPopupValid.php".to_string(),
-        ".vspscc".to_string(),
-        ".vssscc".to_string(),
-        ".w".to_string(),
-        ".work".to_string(),
-        ".wvx".to_string(),
-        ".xspf".to_string(),
-        ".-".to_string(),
-        ".Admin".to_string(),
-        ".E.".to_string(),
-        ".Engineer".to_string(),
-        ".INC".to_string(),
-        ".LOG.new".to_string(),
-        ".MAXIMIZE".to_string(),
-        ".MPG".to_string(),
-        ".NDM".to_string(),
-        ".Php".to_string(),
-        ".R".to_string(),
-        ".SIM".to_string(),
-        ".SQL".to_string(),
-        ".Services".to_string(),
-        ".[file".to_string(),
-        ".accdb".to_string(),
-        ".act".to_string(),
-        ".actions.php".to_string(),
-        ".admin.php".to_string(),
-        ".ads".to_string(),
-        ".alhtm".to_string(),
-        ".all".to_string(),
-        ".ani".to_string(),
-        ".apf".to_string(),
-        ".apj".to_string(),
-        ".ar".to_string(),
-        ".aral-design.com".to_string(),
-        ".aral-design.de".to_string(),
-        ".arc".to_string(),
-        ".array-key-exists".to_string(),
-        ".asp.old".to_string(),
-        ".asp1".to_string(),
-        ".aspg".to_string(),
-        ".bfhtm".to_string(),
-        ".biminifinder".to_string(),
-        ".br".to_string(),
-        ".browser".to_string(),
-        ".build".to_string(),
-        ".buscar".to_string(),
-        ".categorias".to_string(),
-        ".categories".to_string(),
-        ".ccs".to_string(),
-        ".ch".to_string(),
-        ".cl".to_string(),
-        ".click.php".to_string(),
-        ".cls".to_string(),
-        ".cls.php".to_string(),
-        ".cms.ad.AdServer.cls".to_string(),
-        ".com-tov.html".to_string(),
-        ".com.ar".to_string(),
-        ".com.br".to_string(),
-        ".com.htm".to_string(),
-        ".com.old".to_string(),
-        ".common".to_string(),
-        ".conf.php".to_string(),
-        ".contact.php".to_string(),
-        ".control".to_string(),
-        ".core.php".to_string(),
-        ".counter.php".to_string(),
-        ".coverfinder".to_string(),
-        ".create.php".to_string(),
-        ".cs2".to_string(),
-        ".d2w".to_string(),
-        ".dbm".to_string(),
-        ".dct".to_string(),
-        ".dmb".to_string(),
-        ".doc.doc".to_string(),
-        ".dxf".to_string(),
-        ".ed".to_string(),
-        ".email.shtml".to_string(),
-        ".en.htm".to_string(),
-        ".engine".to_string(),
-        ".env".to_string(),
-        ".error-log".to_string(),
-        ".esp".to_string(),
-        ".ex".to_string(),
-        ".exc".to_string(),
-        ".exe,".to_string(),
-        ".ext".to_string(),
-        ".external".to_string(),
-        ".ficheros".to_string(),
-        ".fichiers".to_string(),
-        ".flush".to_string(),
-        ".fmt".to_string(),
-        ".fn".to_string(),
-        ".footer".to_string(),
-        ".form_jhtml".to_string(),
-        ".friend".to_string(),
-        ".g.".to_string(),
-        ".geo.xml".to_string(),
-        ".ghtml".to_string(),
-        ".google.com".to_string(),
-        ".gov".to_string(),
-        ".gpg".to_string(),
-        ".hl".to_string(),
-        ".href".to_string(),
-        ".htm.d".to_string(),
-        ".htm.html".to_string(),
-        ".htm.old".to_string(),
-        ".htm2".to_string(),
-        ".html.orig".to_string(),
-        ".html.sav".to_string(),
-        ".html[".to_string(),
-        ".html]".to_string(),
-        ".html_".to_string(),
-        ".html_files".to_string(),
-        ".htmlpar".to_string(),
-        ".htmlprint".to_string(),
-        ".html}".to_string(),
-        ".htm~".to_string(),
-        ".hts".to_string(),
-        ".hu".to_string(),
-        ".hwp".to_string(),
-        ".ibf".to_string(),
-        ".il".to_string(),
-        ".image.php".to_string(),
-        ".imagecreatetruecolor".to_string(),
-        ".imagejpeg".to_string(),
-        ".iml".to_string(),
-        ".imprimer".to_string(),
-        ".imprimer-cadre".to_string(),
-        ".imprimir".to_string(),
-        ".imprimir-marco".to_string(),
-        ".info.html".to_string(),
-        ".info.php".to_string(),
-        ".ini.bak".to_string(),
-        ".ini.default".to_string(),
-        ".inl".to_string(),
-        ".inv".to_string(),
-        ".join".to_string(),
-        ".jpg.jpg".to_string(),
-        ".jps".to_string(),
-        ".key".to_string(),
-        ".kit".to_string(),
-        ".lang".to_string(),
-        ".lignee".to_string(),
-        ".ltr".to_string(),
-        ".lzh".to_string(),
-        ".m4a".to_string(),
-        ".mail".to_string(),
-        ".manager".to_string(),
-        ".md5".to_string(),
-        ".met".to_string(),
-        ".metadesc".to_string(),
-        ".metakeys".to_string(),
-        ".mht".to_string(),
-        ".min".to_string(),
-        ".mld".to_string(),
-        ".mobi".to_string(),
-        ".mobile".to_string(),
-        ".mv4".to_string(),
-        ".n".to_string(),
-        ".net-tov.html".to_string(),
-        ".nfo".to_string(),
-        ".nikon".to_string(),
-        ".nodos".to_string(),
-        ".nxg".to_string(),
-        ".obyx".to_string(),
-        ".ods".to_string(),
-        ".old.2".to_string(),
-        ".old.asp".to_string(),
-        ".old.html".to_string(),
-        ".open".to_string(),
-        ".opml.config".to_string(),
-        ".ord".to_string(),
-        ".org.zip".to_string(),
-        ".ori".to_string(),
-        ".partfinder".to_string(),
-        ".pho".to_string(),
-        ".php-".to_string(),
-        ".phpl".to_string(),
-        ".phpx".to_string(),
-        ".pix".to_string(),
-        ".pls".to_string(),
-        ".prc".to_string(),
-        ".pre".to_string(),
-        ".prhtm".to_string(),
-        ".print-frame".to_string(),
-        ".print.".to_string(),
-        ".print.shtml".to_string(),
-        ".printer".to_string(),
-        ".properties".to_string(),
-        ".propfinder".to_string(),
-        ".pvx".to_string(),
-        ".recherche".to_string(),
-        ".redirect".to_string(),
-        ".req".to_string(),
-        ".roshani-gunewardene.net".to_string(),
-        ".roshani-m-gunewardene.com".to_string(),
-        ".safe".to_string(),
-        ".sbk".to_string(),
-        ".se.php".to_string(),
-        ".search.asp".to_string(),
-        ".sec".to_string(),
-        ".seo".to_string(),
-        ".serv".to_string(),
-        ".server.php".to_string(),
-        ".servlet".to_string(),
-        ".settings".to_string(),
-        ".sf".to_string(),
-        ".shopping_return.php".to_string(),
-        ".shopping_return_adsense.php".to_string(),
-        ".show".to_string(),
-        ".sht".to_string(),
-        ".skins".to_string(),
-        ".so".to_string(),
-        ".sph".to_string(),
-        ".split".to_string(),
-        ".sso".to_string(),
-        ".stats.php".to_string(),
-        ".story".to_string(),
-        ".swd".to_string(),
-        ".swf.html".to_string(),
-        ".sys".to_string(),
-        ".tex".to_string(),
-        ".tga".to_string(),
-        ".thm".to_string(),
-        ".tlp".to_string(),
-        ".tml".to_string(),
-        ".tmp.php".to_string(),
-        ".touch".to_string(),
-        ".tsv".to_string(),
-        ".txt.".to_string(),
-        ".txt.html".to_string(),
-        ".ug".to_string(),
-        ".unternehmen".to_string(),
-        ".utf8".to_string(),
-        ".vbproj.vspscc".to_string(),
-        ".vsprintf".to_string(),
-        ".vstemplate".to_string(),
-        ".vtl".to_string(),
-        ".wbmp".to_string(),
-        ".webc".to_string(),
-        ".webproj".to_string(),
-        ".wihtm".to_string(),
-        ".wp".to_string(),
-        ".wps".to_string(),
-        ".wri".to_string(),
-        ".wsc".to_string(),
-        ".www".to_string(),
-        ".xsp".to_string(),
-        ".xsql".to_string(),
-        ".zip,".to_string(),
-        ".zml".to_string(),
-        ".ztml".to_string(),
-        ". EXTRAHOTELERO HOSPEDAJE".to_string(),
-        ". T.".to_string(),
-        ". php".to_string(),
-        ".,".to_string(),
-        ".a5w".to_string(),
-        ".aac".to_string(),
-        ".access".to_string(),
-        ".act.php".to_string(),
-        ".action.php".to_string(),
-        ".actions".to_string(),
-        ".activate.php".to_string(),
-        ".ad.php".to_string(),
-        ".add.php".to_string(),
-        ".adenaw.com".to_string(),
-        ".adm".to_string(),
-        ".advsearch".to_string(),
-        ".ag.php".to_string(),
-        ".aj_".to_string(),
-        ".all.hawaii".to_string(),
-        ".amaphun.com".to_string(),
-        ".andriy.lviv.ua".to_string(),
-        ".ap".to_string(),
-        ".api".to_string(),
-        ".apk".to_string(),
-        ".application".to_string(),
-        ".archiv".to_string(),
-        ".arj".to_string(),
-        ".array-map".to_string(),
-        ".array-values".to_string(),
-        ".art".to_string(),
-        ".artdeco".to_string(),
-        ".articlePk".to_string(),
-        ".artnet.".to_string(),
-        ".ascx.resx".to_string(),
-        ".asia".to_string(),
-        ".asp-".to_string(),
-        ".asp.LCK".to_string(),
-        ".asp.html".to_string(),
-        ".asp2".to_string(),
-        ".aspDONOTUSE".to_string(),
-        ".asp_".to_string(),
-        ".asp_files".to_string(),
-        ".aspl".to_string(),
-        ".aspp".to_string(),
-        ".asps".to_string(),
-        ".aspx.designer.cs".to_string(),
-        ".aspx_files".to_string(),
-        ".aspxx".to_string(),
-        ".aspy".to_string(),
-        ".asxp".to_string(),
-        ".at.html".to_string(),
-        ".avatar.php".to_string(),
-        ".awstats".to_string(),
-        ".babymhiasexy.com".to_string(),
-        ".backup.php".to_string(),
-        ".bak.php".to_string(),
-        ".banan.se".to_string(),
-        ".banner.php".to_string(),
-        ".barnes".to_string(),
-        ".basicmap.php".to_string(),
-        ".baut".to_string(),
-        ".bc".to_string(),
-        ".best-vpn.com".to_string(),
-        ".beta".to_string(),
-        ".biz".to_string(),
-        ".blackandmature.com".to_string(),
-        ".bmp.php".to_string(),
-        ".board.asd".to_string(),
-        ".boom".to_string(),
-        ".bossspy.org".to_string(),
-        ".buscadorpornoxxx.com".to_string(),
-        ".buy-here.com".to_string(),
-        ".buyadspace".to_string(),
-        ".bycategory".to_string(),
-        ".bylocation".to_string(),
-        ".bz".to_string(),
-        ".c.html".to_string(),
-        ".cache.inc.php".to_string(),
-        ".cache.php".to_string(),
-        ".car".to_string(),
-        ".cascinaamalia.it".to_string(),
-        ".cat.php".to_string(),
-        ".catalog".to_string(),
-        ".cdf".to_string(),
-        ".ce".to_string(),
-        ".cfm.bak".to_string(),
-        ".cfsifatest.co.uk".to_string(),
-        ".cfstest.co.uk".to_string(),
-        ".cfswf".to_string(),
-        ".cfx".to_string(),
-        ".cgis".to_string(),
-        ".chat".to_string(),
-        ".chdir".to_string(),
-        ".chloesworld.com".to_string(),
-        ".classes.php".to_string(),
-        ".cmp".to_string(),
-        ".cnt".to_string(),
-        ".co".to_string(),
-        ".co-operativebank.co.uk".to_string(),
-        ".co-operativebanktest.co.uk".to_string(),
-        ".co-operativeinsurance.co.uk".to_string(),
-        ".co-operativeinsurancetest.co.uk".to_string(),
-        ".co-operativeinvestmentstest.co.uk".to_string(),
-        ".co.il".to_string(),
-        ".colorbox-min.js".to_string(),
-        ".com-authorization-required.html".to_string(),
-        ".com-bad-request.html".to_string(),
-        ".com-forbidden.html".to_string(),
-        ".com-internal-server-error.html".to_string(),
-        ".com-page-not-found.html".to_string(),
-        ".com.au".to_string(),
-        ".com.php".to_string(),
-        ".com.ua".to_string(),
-        ".com_Backup_".to_string(),
-        ".com_files".to_string(),
-        ".comments".to_string(),
-        ".comments.".to_string(),
-        ".comments.php".to_string(),
-        ".compiler.php".to_string(),
-        ".conf.html".to_string(),
-        ".confirm.email".to_string(),
-        ".connect.php".to_string(),
-        ".console".to_string(),
-        ".contact".to_string(),
-        ".content.php".to_string(),
-        ".controller".to_string(),
-        ".controls-3.1.5.swf".to_string(),
-        ".cookie.js".to_string(),
-        ".corp".to_string(),
-        ".corp.footer".to_string(),
-        ".cqs".to_string(),
-        ".cron".to_string(),
-        ".cropcanvas.php".to_string(),
-        ".cropinterface.php".to_string(),
-        ".crx".to_string(),
-        ".csproj.webinfo".to_string(),
-        ".csr".to_string(),
-        ".css.LCK".to_string(),
-        ".css.gz".to_string(),
-        ".cssd".to_string(),
-        ".csv.php".to_string(),
-        ".ctp".to_string(),
-        ".cx".to_string(),
-        ".cycle.all.min.js".to_string(),
-        ".d64".to_string(),
-        ".daisy".to_string(),
-        ".dal".to_string(),
-        ".daniel".to_string(),
-        ".daniel-sebald.de".to_string(),
-        ".data.php".to_string(),
-        ".data_".to_string(),
-        ".davis".to_string(),
-        ".dbml".to_string(),
-        ".dcf".to_string(),
-        ".de.jsp".to_string(),
-        ".default.php".to_string(),
-        ".del".to_string(),
-        ".deleted".to_string(),
-        ".dell".to_string(),
-        ".demo".to_string(),
-        ".desarrollo.aquihaydominios.com".to_string(),
-        ".dev.bka.co.nz".to_string(),
-        ".development".to_string(),
-        ".dig".to_string(),
-        ".display.php".to_string(),
-        ".dist".to_string(),
-        ".dk".to_string(),
-        ".dm".to_string(),
-        ".dmca-sucks.com".to_string(),
-        ".dms".to_string(),
-        ".dnn".to_string(),
-        ".dogpl".to_string(),
-        ".donothiredandobrin.com".to_string(),
-        ".dontcopy".to_string(),
-        ".downloadfreeporn.asia".to_string(),
-        ".du".to_string(),
-        ".dump".to_string(),
-        ".dws".to_string(),
-        ".dyn".to_string(),
-        ".ea3ny.com".to_string(),
-        ".easing.min.js".to_string(),
-        ".ebay".to_string(),
-        ".ebay.results.html".to_string(),
-        ".editingoffice.com".to_string(),
-        ".efacil.com.br".to_string(),
-        ".ehtml".to_string(),
-        ".emaximinternational.com".to_string(),
-        ".en.jsp".to_string(),
-        ".enn".to_string(),
-        ".equonix.com".to_string(),
-        ".es.html".to_string(),
-        ".es.jsp".to_string(),
-        ".euforyou.net".to_string(),
-        ".eur".to_string(),
-        ".excel.xml.php".to_string(),
-        ".exec".to_string(),
-        ".exp".to_string(),
-        ".f.l.".to_string(),
-        ".faucetdepot".to_string(),
-        ".faucetdepot.com.vbproj".to_string(),
-        ".faucetdepot.com.vbproj.webinfo".to_string(),
-        ".fb2".to_string(),
-        ".fdml".to_string(),
-        ".feeds.php".to_string(),
-        ".ffa".to_string(),
-        ".ficken.cx".to_string(),
-        ".filereader".to_string(),
-        ".filters.php".to_string(),
-        ".flac".to_string(),
-        ".flypage".to_string(),
-        ".fon".to_string(),
-        ".forget.pass".to_string(),
-        ".form.php".to_string(),
-        ".forms".to_string(),
-        ".forum".to_string(),
-        ".found".to_string(),
-        ".fp7".to_string(),
-        ".fr.jsp".to_string(),
-        ".freeasianporn.asia".to_string(),
-        ".freepornxxx.asia".to_string(),
-        ".frk".to_string(),
-        ".frontpage.php".to_string(),
-        ".ft".to_string(),
-        ".ftl".to_string(),
-        ".fucks.nl".to_string(),
-        ".funzz.fr".to_string(),
-        ".gallery.php".to_string(),
-        ".garcia".to_string(),
-        ".gb".to_string(),
-        ".get".to_string(),
-        ".get-meta-tags".to_string(),
-        ".gif         ".to_string(),
-        ".gif.count".to_string(),
-        ".girlvandiesuburbs.co.za".to_string(),
-        ".gitihost.com".to_string(),
-        ".glasner.ru".to_string(),
-        ".google".to_string(),
-        ".gray".to_string(),
-        ".gsp".to_string(),
-        ".guiaweb.tk".to_string(),
-        ".gutschein".to_string(),
-        ".guy".to_string(),
-        ".ha".to_string(),
-        ".hardestlist.com".to_string(),
-        ".hardpussy.com".to_string(),
-        ".hasrett.de".to_string(),
-        ".hawaii".to_string(),
-        ".header.php".to_string(),
-        ".henry".to_string(),
-        ".him".to_string(),
-        ".history".to_string(),
-        ".hlr".to_string(),
-        ".hm".to_string(),
-        ".ho".to_string(),
-        ".hokkaido".to_string(),
-        ".hold".to_string(),
-        ".home.php".to_string(),
-        ".home.test".to_string(),
-        ".homepage".to_string(),
-        ".hp".to_string(),
-        ".htm.bak".to_string(),
-        ".htm.rc".to_string(),
-        ".htm3".to_string(),
-        ".htm5".to_string(),
-        ".htm7".to_string(),
-        ".htm8".to_string(),
-        ".htm_".to_string(),
-        ".html,,".to_string(),
-        ".html-0".to_string(),
-        ".html-1".to_string(),
-        ".html-c".to_string(),
-        ".html-old".to_string(),
-        ".html-p".to_string(),
-        ".html.htm".to_string(),
-        ".html.images".to_string(),
-        ".html.inc".to_string(),
-        ".html.none".to_string(),
-        ".html.pdf".to_string(),
-        ".html.start".to_string(),
-        ".html.txt".to_string(),
-        ".html4".to_string(),
-        ".html5".to_string(),
-        ".html7".to_string(),
-        ".htmlBAK".to_string(),
-        ".htmlDolmetschen".to_string(),
-        ".html_old".to_string(),
-        ".htmla".to_string(),
-        ".htmlc".to_string(),
-        ".htmlfeed".to_string(),
-        ".htmlq".to_string(),
-        ".htmlu".to_string(),
-        ".htn".to_string(),
-        ".htpasswd".to_string(),
-        ".iac.".to_string(),
-        ".ibuysss.info".to_string(),
-        ".iconv".to_string(),
-        ".idf".to_string(),
-        ".iframe_filtros".to_string(),
-        ".ignore.php".to_string(),
-        ".ihmtl".to_string(),
-        ".ihya".to_string(),
-        ".imp".to_string(),
-        ".in".to_string(),
-        ".inactive".to_string(),
-        ".inc.php.bak".to_string(),
-        ".inc.php3".to_string(),
-        ".incest-porn.sex-startje.nl".to_string(),
-        ".incestporn.sex-startje.nl".to_string(),
-        ".incl".to_string(),
-        ".indiansexzite.com".to_string(),
-        ".indt".to_string(),
-        ".ini.NEWCONFIGPOSSIBLYBROKEN".to_string(),
-        ".insert".to_string(),
-        ".internet-taxprep.com".to_string(),
-        ".interpreterukraine.com".to_string(),
-        ".ipl".to_string(),
-        ".issues".to_string(),
-        ".itml".to_string(),
-        ".ixi".to_string(),
-        ".jhtm".to_string(),
-        ".job".to_string(),
-        ".joseph".to_string(),
-        ".jpf".to_string(),
-        ".jpg.xml".to_string(),
-        ".jpg[".to_string(),
-        ".jpg]".to_string(),
-        ".js,".to_string(),
-        ".js.LCK".to_string(),
-        ".jsa".to_string(),
-        ".jsd".to_string(),
-        ".jso".to_string(),
-        ".jsp.old".to_string(),
-        ".jsps".to_string(),
-        ".jtp".to_string(),
-        ".keyword".to_string(),
-        ".kinkywear.net".to_string(),
-        ".kk".to_string(),
-        ".knvbcommunicator.voetbalassist.nl".to_string(),
-        ".kokuken".to_string(),
-        ".ks".to_string(),
-        ".kutxa.net-en".to_string(),
-        ".lang-de.php".to_string(),
-        ".lang.php".to_string(),
-        ".langhampartners.com".to_string(),
-        ".lappgroup.com".to_string(),
-        ".last".to_string(),
-        ".latest".to_string(),
-        ".lha".to_string(),
-        ".links".to_string(),
-        ".list.includes".to_string(),
-        ".listMiniGrid".to_string(),
-        ".listing".to_string(),
-        ".lng".to_string(),
-        ".loc".to_string(),
-        ".local.cfm".to_string(),
-        ".location.href".to_string(),
-        ".log2".to_string(),
-        ".lua".to_string(),
-        ".lynkx".to_string(),
-        ".maastrichtairporthotels.com".to_string(),
-        ".mag".to_string(),
-        ".mail.php".to_string(),
-        ".malesextoys.us".to_string(),
-        ".massivewankers.com".to_string(),
-        ".mbizgroup".to_string(),
-        ".mel".to_string(),
-        ".members".to_string(),
-        ".meretrizdelujo.com".to_string(),
-        ".messagey.com".to_string(),
-        ".metadata.js".to_string(),
-        ".meus.php".to_string(),
-        ".midi".to_string(),
-        ".milliculture.net".to_string(),
-        ".min_".to_string(),
-        ".miss-video.com".to_string(),
-        ".mk.gutschein".to_string(),
-        ".mk.rabattlp".to_string(),
-        ".mkv".to_string(),
-        ".mmap".to_string(),
-        ".model-escorts.asia".to_string(),
-        ".modelescorts.asia".to_string(),
-        ".mp".to_string(),
-        ".mp3.html".to_string(),
-        ".mq4".to_string(),
-        ".mreply.rc".to_string(),
-        ".msp".to_string(),
-        ".mvn".to_string(),
-        ".mysqli".to_string(),
-        ".napravlenie_ASC".to_string(),
-        ".napravlenie_DESC".to_string(),
-        ".nded-pga-emial".to_string(),
-        ".net-en".to_string(),
-        ".net-print.htm".to_string(),
-        ".net_Backup_Giornaliero".to_string(),
-        ".net_Backup_Settimanale".to_string(),
-        ".new.htm".to_string(),
-        ".newsletter".to_string(),
-        ".nexucom.com".to_string(),
-        ".ninwinter.net".to_string(),
-        ".nl.html".to_string(),
-        ".nonude.org".to_string(),
-        ".nonudes.com".to_string(),
-        ".nth".to_string(),
-        ".nz".to_string(),
-        ".od".to_string(),
-        ".offer.php".to_string(),
-        ".offline".to_string(),
-        ".ogv".to_string(),
-        ".ok".to_string(),
-        ".old.1".to_string(),
-        ".old.htm".to_string(),
-        ".old.old".to_string(),
-        ".old1".to_string(),
-        ".old3".to_string(),
-        ".older".to_string(),
-        ".oliver".to_string(),
-        ".onedigitalcentral.com".to_string(),
-        ".onenettv.com".to_string(),
-        ".online".to_string(),
-        ".opensearch".to_string(),
-        ".org-tov.html".to_string(),
-        ".org.ua-tov.html".to_string(),
-        ".orig.html".to_string(),
-        ".origin.php".to_string(),
-        ".original.html".to_string(),
-        ".orlando-vacationhome.net".to_string(),
-        ".orlando-vacationhomes-pools.com".to_string(),
-        ".orlando-vacationrentals.net".to_string(),
-        ".osg".to_string(),
-        ".outbound".to_string(),
-        ".owen".to_string(),
-        ".ownhometest.co.uk".to_string(),
-        ".pae".to_string(),
-        ".page_pls_all_password".to_string(),
-        ".pages-medicales.com".to_string(),
-        ".pan".to_string(),
-        ".parse-url".to_string(),
-        ".part".to_string(),
-        ".pass".to_string(),
-        ".patch".to_string(),
-        ".paul".to_string(),
-        ".paymethods.php".to_string(),
-        ".pazderski.com".to_string(),
-        ".pazderski.net".to_string(),
-        ".pazderski.us".to_string(),
-        ".pdd".to_string(),
-        ".pdf.html".to_string(),
-        ".pdf.pdf".to_string(),
-        ".pdf.php".to_string(),
-        ".pdfx".to_string(),
-        ".perfect-color-world.com".to_string(),
-        ".petersburg-apartments-for-business.html".to_string(),
-        ".petersburg-apartments-for-tourists.html".to_string(),
-        ".petersburg-romantic-apartments.html".to_string(),
-        ".phdo".to_string(),
-        ".photo".to_string(),
-        ".php--------------".to_string(),
-        ".php.LCK".to_string(),
-        ".php.backup".to_string(),
-        ".php.html".to_string(),
-        ".php.inc".to_string(),
-        ".php.mno".to_string(),
-        ".php.original".to_string(),
-        ".php_".to_string(),
-        ".php_OLD".to_string(),
-        ".php_old".to_string(),
-        ".phphp".to_string(),
-        ".phppar".to_string(),
-        ".phpvreor.php".to_string(),
-        ".php".to_string(),
-        ".pht".to_string(),
-        ".pl.html".to_string(),
-        ".planetcom.ca".to_string(),
-        ".playwithparis.com".to_string(),
-        ".plugins".to_string(),
-        ".png,bmp".to_string(),
-        ".popup".to_string(),
-        ".pornfailures.com".to_string(),
-        ".pornoizlee.tk".to_string(),
-        ".pornz.tv".to_string(),
-        ".posting.prep".to_string(),
-        ".prev".to_string(),
-        ".print.jsp".to_string(),
-        ".prl".to_string(),
-        ".prosdo.com".to_string(),
-        ".psb".to_string(),
-        ".publisher.php".to_string(),
-        ".puresolo.com".to_string(),
-        ".pussyjourney.com".to_string(),
-        ".qtgp".to_string(),
-        ".qxd".to_string(),
-        ".r.".to_string(),
-        ".rabattlp".to_string(),
-        ".rails".to_string(),
-        ".randomocityproductions.com".to_string(),
-        ".rateart.php".to_string(),
-        ".readfile".to_string(),
-        ".rec.html".to_string(),
-        ".redirect.php".to_string(),
-        ".remove".to_string(),
-        ".remove.php".to_string(),
-        ".removed".to_string(),
-        ".resultados".to_string(),
-        ".resume".to_string(),
-        ".rhtm".to_string(),
-        ".riddlesintime.com".to_string(),
-        ".rmvb".to_string(),
-        ".ro".to_string(),
-        ".roma".to_string(),
-        ".roomscity.com".to_string(),
-        ".roshanigunewardene.com".to_string(),
-        ".rpt".to_string(),
-        ".rsp".to_string(),
-        ".rss.php".to_string(),
-        ".rss_cars".to_string(),
-        ".rss_homes".to_string(),
-        ".rss_jobs".to_string(),
-        ".rtfd".to_string(),
-        ".rvt".to_string(),
-        ".s.html".to_string(),
-        ".sadopasion.com".to_string(),
-        ".safariextz".to_string(),
-        ".salestax.php".to_string(),
-        ".sc".to_string(),
-        ".sca-tork.com".to_string(),
-        ".scandir".to_string(),
-        ".scrollTo.js".to_string(),
-        ".search.html".to_string(),
-        ".sec.cfm".to_string(),
-        ".section".to_string(),
-        ".secure".to_string(),
-        ".send".to_string(),
-        ".sent-".to_string(),
-        ".service".to_string(),
-        ".session-regenerate-id".to_string(),
-        ".set".to_string(),
-        ".sex-startje.nl".to_string(),
-        ".sexmeme.com".to_string(),
-        ".sexon.com".to_string(),
-        ".sexy-girls4abo.de".to_string(),
-        ".sfw".to_string(),
-        ".sgf".to_string(),
-        ".shipcode.php".to_string(),
-        ".shipdiscount.php".to_string(),
-        ".show.php".to_string(),
-        ".shtml.html".to_string(),
-        ".sidebar".to_string(),
-        ".sisx".to_string(),
-        ".sitemap.".to_string(),
-        ".skin".to_string(),
-        ".small-penis-humiliation.net".to_string(),
-        ".smiletest.co.uk".to_string(),
-        ".snippet.aspx".to_string(),
-        ".snuffx.com".to_string(),
-        ".sort".to_string(),
-        ".sortirovka_Price.napravlenie_ASC".to_string(),
-        ".sortirovka_Price.napravlenie_DESC".to_string(),
-        ".sortirovka_customers_rating.napravlenie_ASC".to_string(),
-        ".sortirovka_customers_rating.napravlenie_DESC".to_string(),
-        ".sortirovka_name.napravlenie_ASC".to_string(),
-        ".sortirovka_name.napravlenie_DESC".to_string(),
-        ".sp".to_string(),
-        ".sphp3".to_string(),
-        ".srch".to_string(),
-        ".srf".to_string(),
-        ".srvl".to_string(),
-        ".st-patricks.com".to_string(),
-        ".sta".to_string(),
-        ".staged.php".to_string(),
-        ".staging".to_string(),
-        ".start.php".to_string(),
-        ".stat".to_string(),
-        ".stats".to_string(),
-        ".step".to_string(),
-        ".stml".to_string(),
-        ".storebanner.php".to_string(),
-        ".storelogo.php".to_string(),
-        ".storename.php".to_string(),
-        ".sts.php".to_string(),
-        ".suarez".to_string(),
-        ".submit".to_string(),
-        ".support".to_string(),
-        ".support.html".to_string(),
-        ".swf.LCK".to_string(),
-        ".sym".to_string(),
-        ".system".to_string(),
-        ".tab-".to_string(),
-        ".table.html".to_string(),
-        ".tablesorter.min.js".to_string(),
-        ".tablesorter.pager.js".to_string(),
-        ".tatianyc.com".to_string(),
-        ".tb".to_string(),
-        ".tech".to_string(),
-        ".teen-shy.com".to_string(),
-        ".teenhardpussy.com".to_string(),
-        ".temp.php".to_string(),
-        ".templates.php".to_string(),
-        ".temporarily.withdrawn.html".to_string(),
-        ".test.cgi".to_string(),
-        ".test.php".to_string(),
-        ".tf".to_string(),
-        ".tg".to_string(),
-        ".thanks".to_string(),
-        ".thehotfish.com".to_string(),
-        ".theme".to_string(),
-        ".thompson".to_string(),
-        ".thumb.jpg".to_string(),
-        ".ticket.submit".to_string(),
-        ".tim".to_string(),
-        ".tk".to_string(),
-        ".tls".to_string(),
-        ".to".to_string(),
-        ".touch.action".to_string(),
-        ".trace".to_string(),
-        ".tracker.ashx".to_string(),
-        ".trade".to_string(),
-        ".trishasex.viedos.com".to_string(),
-        ".ts".to_string(),
-        ".tst".to_string(),
-        ".tvpi".to_string(),
-        ".txt.txt".to_string(),
-        ".txuri-urdin.com".to_string(),
-        ".ufo".to_string(),
-        ".ugmart.ug".to_string(),
-        ".ui-1.5.2".to_string(),
-        ".unixteacher.org".to_string(),
-        ".unsharp.php".to_string(),
-        ".update".to_string(),
-        ".upgrade".to_string(),
-        ".v1.11.js".to_string(),
-        ".v2.php".to_string(),
-        ".vacationhomes-pools.com".to_string(),
-        ".var".to_string(),
-        ".venetian.com,prod2.venetian.com,reservations.venetian.com,".to_string(),
-        ".verify".to_string(),
-        ".video".to_string(),
-        ".videodeputas.com".to_string(),
-        ".videos-chaudes.com".to_string(),
-        ".viewpage__10".to_string(),
-        ".vmdk".to_string(),
-        ".vn".to_string(),
-        ".voetbalassist.nl".to_string(),
-        ".vs".to_string(),
-        ".vx".to_string(),
-        ".vxlpub".to_string(),
-        ".w3m".to_string(),
-        ".w3x".to_string(),
-        ".wax".to_string(),
-        ".web-teck.com".to_string(),
-        ".webalizer".to_string(),
-        ".webarchive".to_string(),
-        ".webjockey.nl".to_string(),
-        ".webm".to_string(),
-        ".weedooz.eu".to_string(),
-        ".wgx".to_string(),
-        ".wimzi.php".to_string(),
-        ".wireless".to_string(),
-        ".wireless.action".to_string(),
-        ".wm".to_string(),
-        ".woolovers.com".to_string(),
-        ".working".to_string(),
-        ".wpl".to_string(),
-        ".wplus".to_string(),
-        ".wps.rtf".to_string(),
-        ".write.php".to_string(),
-        ".wwsec_app_priv.login".to_string(),
-        ".www.annuaire-vimarty.net".to_string(),
-        ".www.annuaire-web.info".to_string(),
-        ".www.kit-graphik.com".to_string(),
-        ".www.photo-scope.fr".to_string(),
-        ".xcam.at".to_string(),
-        ".xconf".to_string(),
-        ".xcwc.com".to_string(),
-        ".xgi".to_string(),
-        ".xhtml5".to_string(),
-        ".xlt".to_string(),
-        ".xm".to_string(),
-        ".xml.old".to_string(),
-        ".xpdf".to_string(),
-        ".xqy".to_string(),
-        ".xslx".to_string(),
-        ".xst".to_string(),
-        ".xsx".to_string(),
-        ".xy.php".to_string(),
-        ".yp".to_string(),
-        ".ys".to_string(),
-        ".z".to_string(),
-        ".za".to_string(),
-        ".zh.html".to_string(),
-        ".zhtml".to_string(),
-        ".zip.php".to_string(),
-    ];
-    valid_extension.contains(&tested_extension.to_string())
-}
-
-pub fn remove_extension(file_name: &str) -> String {
-    let extension = get_extension(file_name);
-    if extension.is_none() {
-        return file_name.to_owned();
-    }
-
-    let regex_string = format!(r"\.{}$", extension.unwrap());
-    let extension_regex = Regex::new(&regex_string).unwrap();
-    let removed = extension_regex.replace_all(file_name, "");
-
-    format!("{}", removed)
-}
-
-pub fn get_extension(file_name: &str) -> Option<String> {
-    let path = Path::new(file_name);
-    let p_extension = path.extension()?;
-    let string_extension = format!(".{}", p_extension.to_str().unwrap());
-    if !is_valid_extension(&string_extension) {
-        return None;
-    }
-    Some(p_extension.to_str().unwrap().to_string())
-}
-
 pub fn remove_ignored_string(working_string: &str, ignored_str: Vec<String>) -> String {
     let mut return_string = working_string.to_string();
     for i_s in ignored_str {
@@ -1813,40 +23,4 @@ pub fn normalize(to_normalize: &str) -> String {
         .cloned()
         .collect::<String>()
         .to_uppercase()
-}
-
-pub fn is_hexa(p_hexa: &str) -> bool {
-    let hexa_char = Regex::new(r"^[0-9A-F]+$").unwrap();
-    hexa_char.is_match(&p_hexa.to_uppercase())
-}
-
-pub fn is_digit(w: &str) -> bool {
-    let digit_regex = Regex::new(r"^[-+0-9A-F]+$").unwrap();
-    digit_regex.is_match(&w.to_uppercase())
-}
-
-pub fn split_by_delimiter(raw: &str, delimiter: Vec<char>) -> Vec<&str> {
-    let mut raw_string = String::default();
-    for s in delimiter {
-        raw_string = format!(r"{}\{}", raw_string, s);
-    }
-    let regex_split_str = format!(r"([{}])", raw_string);
-    let split_regex = Regex::new(&regex_split_str).unwrap();
-    split_regex.split(raw).collect::<Vec<&str>>()
-}
-
-/// private function test
-#[test]
-fn test_valid_extension() {
-    assert!(is_valid_extension(".mkv"));
-    assert!(is_valid_extension(".mp4"));
-    assert!(is_valid_extension(".avi"));
-    assert!(is_valid_extension(".html"));
-}
-
-#[test]
-fn test_invalid_extension() {
-    assert!(!is_valid_extension("mkv"));
-    assert!(!is_valid_extension(".non an extension"));
-    assert!(!is_valid_extension("avi"));
 }
